Backend/requirements.txt
Quart==0.17.0
hypercorn==0.13.2
uvloop==0.16.0
motor==3.0.0
google-cloud-pubsub==2.19.0
google-cloud-storage==2.13.0
imutils==0.5.4
moviepy==1.0.3
nltk==3.8.1
numpy==1.24.3
openai==1.4.0
openai_whisper==20231117
opencv-python==4.8.1.78
pandas==2.0.2
protobuf==4.25.1
python-dotenv==1.0.0
pytube==15.0.0
redis==4.3.4
retina-face==0.0.13
tqdm==4.60.0
yt-dlp==2023.11.16
